fin de cambio 8
todo ok en pdf's

+++ lib/functions/utils/delete_files_in_folder.py
@@ -3,7 +3,7 @@ import shutil
 
 # Borra los archivos temporales
 
-def deleteFilesInFolder(folderPath, fileNamePrefix=None):
+def deleteFilesInFolder(folderPath, fileNamePrefix=None, deleteDirectories=False):
     folder = folderPath
     for filename in os.listdir(folder):
         if fileNamePrefix is None or filename.startswith(fileNamePrefix):
@@ -11,7 +11,7 @@ def deleteFilesInFolder(folderPath, fileNamePrefix=None):
             try:
                 if os.path.isfile(file_path) or os.path.islink(file_path):
                     os.unlink(file_path)
-                elif os.path.isdir(file_path):
+                elif os.path.isdir(file_path) and deleteDirectories:
                     shutil.rmtree(file_path)
             except Exception as e:
                 print("Failed to delete %s. Reason: %s" % (file_path, e))
